% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dlm_tibbles.R
\name{tidy_cumul_fits}
\alias{tidy_cumul_fits}
\title{get tibble of cumulative estimates at each lag}
\usage{
tidy_cumul_fits(cpred)
}
\arguments{
\item{cpred}{a crosspred object created from dlnm::crosspred()}
}
\value{
a tibble containing lag, estimate, se, and lower and upper
95\% confidence bounds
}
\description{
get tibble of cumulative estimates at each lag
}
