% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dlm_tibbles.R
\name{tidy_lag_fits}
\alias{tidy_lag_fits}
\title{get tibble of estimates at each lag}
\usage{
tidy_lag_fits(cpred)
}
\arguments{
\item{cpred}{a crosspred object created from \code{\link[dlnm:crosspred]{dlnm::crosspred()}}}
}
\value{
a tibble with columns for lag, estimate, se, and lower and upper
95\% confidence bounds (one row per lag)
}
\description{
get tibble of estimates at each lag
}
