% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dlm_plots.R
\name{tidy_lag_plot}
\alias{tidy_lag_plot}
\title{plot estimate at each lag and confidence intervals using ggplot2}
\usage{
tidy_lag_plot(
  lag_fits,
  continuous = TRUE,
  shading = FALSE,
  shade_colors = c("red", NA, "blue")
)
}
\arguments{
\item{lag_fits}{tibble containing lag, estimate, and lower and upper bounds of
confidence intervals; most likely output from [\verb{tidydlnm::tidy_lag_fits()``] or [}tidydlnm::tidy_cumul_fits()``]}

\item{continuous}{logical. When TRUE, creates a plot with \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} and \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}
(used when \code{arglag} is a continuous function). When FALSE, creates a plot with
\code{\link[ggplot2:geom_linerange]{ggplot2::geom_pointrange()}} (used when \code{arglag} is not a continuous function -- eg,
'integer' or 'strata')}

\item{shading}{logical. When TRUE, adds colored shading to the regions of the plot
corresponding to significant associations as defined by the signSum
variable of lag_fits}

\item{shade_colors}{vector of three colors desired to shade the regions of negative, null,
and positive association, respectively.}
}
\value{
a ggplot object
}
\description{
plot estimate at each lag and confidence intervals using ggplot2
}
