% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dlm_tibbles.R
\name{tidy_overall_fit}
\alias{tidy_overall_fit}
\title{get tibble of overall effect estimate}
\usage{
tidy_overall_fit(cpred)
}
\arguments{
\item{cpred}{a crosspred object created from dlnm::crosspred()}
}
\value{
a tibble containing estimate, se, and lower and upper
95\% confidence bounds (1 row)
}
\description{
get tibble of overall effect estimate
}
