% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dlm_tibbles.R
\name{tidy_window_summary}
\alias{tidy_window_summary}
\title{get tibble detailing each window of association with summary statistics}
\usage{
tidy_window_summary(lag_fits)
}
\arguments{
\item{lag_fits}{tibble containing lag, estimate, and lower and upper bounds of
confidence intervals; most likely output from tidy_lag_fits() or
tidy_cumul_fits()}
}
\value{
a tibble containing windows defined by starting and ending lags, window length,
maximum absolute effect estimate, and the lag at which that maximum occurs
}
\description{
get tibble detailing each window of association with summary statistics
}
