% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydlnm-package.R
\docType{package}
\name{tidydlnm-package}
\alias{tidydlnm}
\alias{tidydlnm-package}
\title{tidydlnm: Wrapper Package for Tidy DLNM Output and Plots}
\description{
tidydlnm is a wrapper package that converts output from the dlnm package into more tidyverse-friendly tibbles and plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://geomarker.io/tidydlnm/}
}

}
\author{
\strong{Maintainer}: Erika Rasnick \email{erika.rasnick@cchmc.org}

}
\keyword{internal}
