% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dlm_tibbles.R
\name{tidy_cumul_fits}
\alias{tidy_cumul_fits}
\title{get tibble of cumulative estimates at each lag}
\usage{
tidy_cumul_fits(cpred, call = rlang::caller_env())
}
\arguments{
\item{cpred}{a crosspred object created from \code{\link[dlnm:crosspred]{dlnm::crosspred()}}}

\item{call}{used for error handling}
}
\value{
a tibble with columns for lag, estimate, se, and lower and upper
95\% confidence bounds (one row per lag)
}
\description{
get tibble of cumulative estimates at each lag
}
