% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dlm_plots.R
\name{tidy_lag_plot}
\alias{tidy_lag_plot}
\title{plot estimate at each lag and confidence intervals using ggplot2}
\usage{
tidy_lag_plot(lag_fits, continuous = TRUE)
}
\arguments{
\item{lag_fits}{tibble containing lag, estimate, and lower and upper bounds of
confidence intervals; most likely output from tidy_lag_fits() or
tidy_cumul_fits()}

\item{continuous}{logical. When TRUE, creates a plot with geom_line and geom_ribbon (used
when arglag is a continuous function). When FALSE, creates a plot with
geom_pointrange (used when arglag is not a continuous function -- eg,
'integer' or 'strata')}
}
\value{
a ggplot object
}
\description{
plot estimate at each lag and confidence intervals using ggplot2
}
